Frame Puzzle and Image editor
Wrote an extensive, semester-long, 27 class and interface program in
Java that alters pixel colors. I also created a user interface that
incorporates a sliding feature in order to incrementally alter pixels.
Through this I developed a frame puzzle that divides a photo into a 5X5
grid and uses key events to swap frame